+++ src/features/TextToImage/config/exports.ts
@@ -1,3 +1,3 @@
-export * as DynamicConfig from './utilities/DynamicConfig';
+export * as DynamicConfig from './utilities/DynamicConfig/exports';
 export * as StaticConfig from './utilities/StaticConfig';
 export * from './utilities/emptyConfig';
